// generated/src/app/api-models/pmb-tax-exceptions.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PmbTaxExceptions {
  seqTaxExceptionId: number;
  seqGpbilId: number;
  seqGroupId: number;
  procedureName: string;
  tableName: string;
  userName: string;
  errorCode: string;
  errorMessage: string;
  errorDatetime: Date;
  comments: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}